Add files via upload
The prototype on which we were working was "The Smart Life for Blind". Our target users are those who lost their vision due to some accident.
To solve their problems, we built two devices :-
1-)Smart Stick 
This device helps in navigation.
The stick can be activated or deactivated simply on double tap. Once activated, UltraSonic sensors come into play which detects obstacles. SOS button on the top of the stick can send e-mail to relatives and friends in emergency condition.
The mail consist of live location of the blind person along with an IP address for the video streaming of his/her surroundings. The live stream will be stored on the cloud.

2)Smart Dustbin
This device works on voice commands of the user which is connected to his smartphone.
Blind person can say simple phrases like "dustbin come",  "dustbin go" and dustbin will follow those instructions.